# back/swagger-spec.yaml
openapi: 3.0.0
paths:
...
  /cart:
    get:
      operationId: CartController_getCart
      parameters: []
      responses:
        '200':
          description: Carrinho retornado com sucesso
      security: &ref_10
        - bearer: []
      summary: Obter carrinho do cliente
      tags: &ref_11
        - Cart
  /cart/add:
    post:
      operationId: CartController_addToCart
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddToCartDto'
      responses:
        '201':
          description: Produto adicionado com sucesso
        '404':
          description: Produto não encontrado
      security: *ref_10
      summary: Adicionar produto ao carrinho
      tags: *ref_11
  /cart/item/{itemId}:
    patch:
      operationId: CartController_updateCartItem
      parameters:
        - name: itemId
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCartItemDto'
      responses:
        '200':
          description: Item atualizado com sucesso
        '404':
          description: Item não encontrado
      security: *ref_10
      summary: Atualizar quantidade de item no carrinho
      tags: *ref_11
    delete:
      operationId: CartController_removeFromCart
      parameters:
        - name: itemId
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Item removido com sucesso
        '404':
          description: Item não encontrado
      security: *ref_10
      summary: Remover item do carrinho
      tags: *ref_11
  /cart/clear:
    delete:
      operationId: CartController_clearCart
      parameters: []
      responses:
        '200':
          description: Carrinho limpo com sucesso
      security: *ref_10
      summary: Limpar carrinho
      tags: *ref_11
...
info:
  title: Ze da Horta API
  description: API documentation for the NestJS application
  version: '1.0'
  contact: {}
...
servers: []
components:
  securitySchemes:
    JWT-auth:
      scheme: bearer
      bearerFormat: JWT
      type: http
      name: JWT
      description: Insira seu token JWT
      in: header
  schemas:
...
    AddToCartDto:
      type: object
      properties:
        productId:
          type: number
          description: ID do produto a ser adicionado ao carrinho
          example: 1
        quantity:
          type: number
          description: Quantidade do produto
          example: 2
          minimum: 1
      required:
        - productId
        - quantity
    UpdateCartItemDto:
      type: object
      properties:
        quantity:
          type: number
          description: Nova quantidade do item
          example: 3
          minimum: 1
      required:
        - quantity
...